Add (optional) support for the AiiDA Gaussian Datatypes (#104)
This PR adds optional support for the Gaussian Datatypes available
in `aiida-gaussian-datatypes` package. User can decide whether to use
the "standard way" of specifying pseudos/basis sets in the input dictionary or
to let the input plugin automatically pick them up from the input nodes. To learn
how to use the new functionality a user might want to have a look in the 
 `examples/gaussian_datatypes` folder.