`Cp2kBaseWorkChain`: Fix restart for no `output_structure` (#151)
For some CP2K calculations, e.g. a "single-point" calculation, there is
no `output_structure` output in the `Cp2kCalculation`. In this case the
current `Cp2KBaseWorkChain` fails to restart, since the
`overwrite_input_structure` step in the outline _expects_ this output to
be present.

Here we simply add a check to see if `output_structure` is one of the
outputs of the final `Cp2kCalculation`, and only update the input
structure in the context if that is the case.