Introduce new `restart_incomplete_calculation` handler, improve parsing (#156)
- Add `_read_stdout` and `_check_stdout_for_errors` methods of the base parser that
are generic and can be reused in all the derived parser. This makes the code shorter.
- Add `add_ext_restart_section` and `add_wfn_restart_section` capable of adding only
the corresponding sections.
- Add a bunch of exit codes to the `Cp2kBaseWorkChain`.
- Introduce a new `restart_incomplete_calculation` handler as a replacement for the
`resubmit_unconverged_geometry` one. Added a deprecation warning in the docstring
of the latter one. Note: every time `restart_incomplete_calculation` detects the ability
to restart a work chain, it launches `add_wfn_restart_section` and (sometimes)
`add_ext_restart_section`. This happens always even if the corresponding sections are
already present in the input dictionary.
- Adapt examples to the changes in the code.